requirements.txt
# ── Core RAG stack ────────────────────────────────────────────
langchain==0.3.25
langchain-openai==0.3.16
langchain-chroma==0.2.4
langchain-community==0.3.24
langgraph==0.4.1
openai==1.77.0
chromadb>=1.0.9
pypdf==5.4.0
python-dotenv==1.1.0
typing-extensions==4.13.2

# ── LangSmith tracing ─────────────────────────────────────────
langsmith==0.3.42

# ── FastAPI REST server ───────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3

# ── Streamlit web UI ──────────────────────────────────────────
streamlit==1.45.0
requests==2.32.3
